rename backup-metadata to backup-metadata-generator
renamed to make it more clear that this component provides metadata for
a backup

[#175080281](https://www.pivotaltracker.com/story/show/175080281)

Authored-by: Tim Downey <tdowney@vmware.com>
