+++ dev-templates/kbld.yml
@@ -28,7 +28,7 @@ sources:
     path: #@ data.values.src_dirs.registry_buddy
     pack: *packConfig
   - imageRepo: cloudfoundry/cf-api-backup-metadata-generator
-    path: #@ data.values.src_dirs.backup_metadata
+    path: #@ data.values.src_dirs.backup_metadata_generator
     pack: &goWithProcfilePackConfig
       build:
         builder: paketobuildpacks/builder:full
@@ -36,8 +36,8 @@ sources:
           - gcr.io/paketo-buildpacks/go
           - gcr.io/paketo-buildpacks/procfile
         rawOptions: ["--default-process", "wait"]
-  - imageRepo: #@ data.values.kbld.destination.backup_metadata
-    path: #@ data.values.src_dirs.backup_metadata
+  - imageRepo: #@ data.values.kbld.destination.backup_metadata_generator
+    path: #@ data.values.src_dirs.backup_metadata_generator
     pack: *goWithProcfilePackConfig
 ---
 apiVersion: kbld.k14s.io/v1alpha1
@@ -56,9 +56,9 @@ destinations:
   - imageRepo: #@ data.values.kbld.destination.registry_buddy
     newImage: #@ data.values.kbld.destination.registry_buddy
   - imageRepo: cloudfoundry/cf-api-backup-metadata-generator
-    newImage: #@ data.values.kbld.destination.backup_metadata
-  - imageRepo: #@ data.values.kbld.destination.backup_metadata
-    newImage: #@ data.values.kbld.destination.backup_metadata
+    newImage: #@ data.values.kbld.destination.backup_metadata_generator
+  - imageRepo: #@ data.values.kbld.destination.backup_metadata_generator
+    newImage: #@ data.values.kbld.destination.backup_metadata_generator
 ---
 apiVersion: kbld.k14s.io/v1alpha1
 kind: ImageKeys
@@ -66,7 +66,7 @@ keys:
 - ccng
 - cf_api_controllers
 - registry_buddy
-- backup_metadata
+- backup_metadata_generator
 ---
 apiVersion: kbld.k14s.io/v1alpha1
 kind: Config
