+++ src/backup-metadata-generator/.gitignore
@@ -1,0 +1,2 @@
+backup-metadata-generator
+
