+++ src/backup-metadata-generator/Procfile
@@ -1,0 +1,2 @@
+wait: sleep infinity
+generate-metadata: backup-metadata-generator
