+++ src/backup-metadata-generator/internal/cfmetadata/cfmetadatafakes/fake_cf_client.go
@@ -4,7 +4,7 @@ package cfmetadatafakes
 import (
 	"sync"
 
-	"code.cloudfoundry.org/capi-k8s-release/src/backup-metadata/internal/cfmetadata"
+	"code.cloudfoundry.org/capi-k8s-release/src/backup-metadata-generator/internal/cfmetadata"
 )
 
 type FakeCfClient struct {
