+++ src/backup-metadata-generator/test/binary/binary_suite_test.go
