+++ src/backup-metadata-generator/test/binary/cf_metadata_compare_test.go
