+++ src/backup-metadata-generator/test/binary/fixtures/cf_state/apps.json
