+++ src/backup-metadata-generator/test/binary/fixtures/cf_state/metadata.json
