+++ src/backup-metadata-generator/test/binary/fixtures/comparer/compare_json/current-minus-app.json
