+++ src/backup-metadata-generator/test/binary/input_parsing_test.go
