+++ src/backup-metadata-generator/test/binary/metadata_retrieval_test.go
