+++ src/backup-metadata-generator/test/e2e/e2e_test.go
@@ -11,7 +11,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"code.cloudfoundry.org/capi-k8s-release/src/backup-metadata/internal/cfmetadata"
+	"code.cloudfoundry.org/capi-k8s-release/src/backup-metadata-generator/internal/cfmetadata"
 )
 
 var _ = Describe("e2e", func() {
