--- src/backup-metadata/.gitignore
@@ -1,2 +1,0 @@
-backup-metadata
-
