+++ values/_default.yml
@@ -61,7 +61,7 @@ images:
   nginx:
   statsd_exporter:
   registry_buddy:
-  backup_metadata:
+  backup_metadata_generator:
 temporary_disable_v2_staging: true
 honeycomb:
   write_key: ""
